updated changes for audit

+++ src/main/java/com/zandroid/volume/VolumeRepository.java
@@ -1,12 +1,9 @@
 package com.zandroid.volume;
 
-import com.zandroid.volume.Volume;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-
-
 @RepositoryRestResource(collectionResourceRel = "volume", path = "volume")
-public interface VolumeRepository extends CrudRepository<Volume, Long> {
+public interface VolumeRepository extends CrudRepository<Volume, Integer> {
 
 }
